docs/coverage/node/html/test/wpt/xhr/progress-events-response-data-gzip.js.gcov.html

<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN">

<html lang="en">

<head>
  <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
  <title>LCOV - lcov.info - test/wpt/xhr/progress-events-response-data-gzip.js</title>
  <link rel="stylesheet" type="text/css" href="../../../gcov.css">
</head>

<body>

  <table width="100%" border=0 cellspacing=0 cellpadding=0>
    <tr><td class="title">LCOV - code coverage report</td></tr>
    <tr><td class="ruler"><img src="../../../glass.png" width=3 height=3 alt=""></td></tr>

    <tr>
      <td width="100%">
        <table cellpadding=1 border=0 width="100%">
          <tr>
            <td width="10%" class="headerItem">Current view:</td>
            <td width="35%" class="headerValue"><a href="../../../index.html">top level</a> - <a href="index.html">test/wpt/xhr</a> - progress-events-response-data-gzip.js<span style="font-size: 80%;"> (source / <a href="progress-events-response-data-gzip.js.func-sort-c.html">functions</a>)</span></td>
            <td width="5%"></td>
            <td width="15%"></td>
            <td width="10%" class="headerCovTableHead">Hit</td>
            <td width="10%" class="headerCovTableHead">Total</td>
            <td width="15%" class="headerCovTableHead">Coverage</td>
          </tr>
          <tr>
            <td class="headerItem">Test:</td>
            <td class="headerValue">lcov.info</td>
            <td></td>
            <td class="headerItem">Lines:</td>
            <td class="headerCovTableEntry">130</td>
            <td class="headerCovTableEntry">137</td>
            <td class="headerCovTableEntryHi">94.9 %</td>
          </tr>
          <tr>
            <td class="headerItem">Date:</td>
            <td class="headerValue">2025-01-18 23:18:38</td>
            <td></td>
            <td class="headerItem">Functions:</td>
            <td class="headerCovTableEntry">3</td>
            <td class="headerCovTableEntry">3</td>
            <td class="headerCovTableEntryHi">100.0 %</td>
          </tr>
          <tr>
            <td></td>
            <td></td>
            <td></td>
            <td class="headerItem">Branches:</td>
            <td class="headerCovTableEntry">5</td>
            <td class="headerCovTableEntry">7</td>
            <td class="headerCovTableEntryLo">71.4 %</td>
          </tr>
          <tr><td><img src="../../../glass.png" width=3 height=3 alt=""></td></tr>
        </table>
      </td>
    </tr>

    <tr><td class="ruler"><img src="../../../glass.png" width=3 height=3 alt=""></td></tr>
  </table>

  <table cellpadding=0 cellspacing=0 border=0>
    <tr>
      <td><br></td>
    </tr>
    <tr>
      <td>
<pre class="sourceHeading">           Branch data     Line data    Source code</pre>
<pre class="source">
<a name="1"><span class="lineNum">       1 </span>           [<span class="branchCov" title="Branch 0 was taken 1 time"> + </span>]:<span class="lineCov">          1 : import assert from 'node:assert/strict';</span></a>
<a name="2"><span class="lineNum">       2 </span>                :<span class="lineCov">          1 : import { XMLHttpRequest } from '../../../lib/whatwg-xhr.js';</span></a>
<a name="3"><span class="lineNum">       3 </span>                :<span class="lineCov">          1 : </span></a>
<a name="4"><span class="lineNum">       4 </span>           [<span class="branchCov" title="Branch 0 was taken 1 time"> + </span>]:<span class="lineCov">          1 : export default (activeURL) =&gt; {</span></a>
<a name="5"><span class="lineNum">       5 </span>                :<span class="lineCov">          1 : </span></a>
<a name="6"><span class="lineNum">       6 </span>                :<span class="lineCov">          1 :         const xhr = new XMLHttpRequest();</span></a>
<a name="7"><span class="lineNum">       7 </span>                :<span class="lineCov">          1 : </span></a>
<a name="8"><span class="lineNum">       8 </span>                :<span class="lineCov">          1 :         let lastTotal;</span></a>
<a name="9"><span class="lineNum">       9 </span>                :<span class="lineCov">          1 :         let lastLoaded = -1;</span></a>
<a name="10"><span class="lineNum">      10 </span>                :<span class="lineCov">          1 : </span></a>
<a name="11"><span class="lineNum">      11 </span>           [<span class="branchCov" title="Branch 0 was taken 1 time"> + </span>]:<span class="lineCov">          1 :         xhr.addEventListener('loadend', (e) =&gt; {</span></a>
<a name="12"><span class="lineNum">      12 </span>                :<span class="lineCov">          1 :                 let len = parseInt(xhr.getResponseHeader('content-length'), 10);</span></a>
<a name="13"><span class="lineNum">      13 </span>                :<span class="lineCov">          1 : </span></a>
<a name="14"><span class="lineNum">      14 </span>                :<span class="lineCov">          1 :                 if(e.lengthComputable){</span></a>
<a name="15"><span class="lineNum">      15 </span>                :<span class="lineCov">          1 :                         assert.strictEqual(e.total,  len, 'event.total is content-length');</span></a>
<a name="16"><span class="lineNum">      16 </span>                :<span class="lineCov">          1 :                         assert.strictEqual(e.loaded, len, 'event.loaded should be content-length at loadend');</span></a>
<a name="17"><span class="lineNum">      17 </span>           [<span class="branchNoCov" title="Branch 0 was not taken"> - </span>]:<span class="lineCov">          1 :                 }</span></a>
<a name="18"><span class="lineNum">      18 </span>                :<span class="lineNoCov">          0 :                 else{</span></a>
<a name="19"><span class="lineNum">      19 </span>                :<span class="lineNoCov">          0 :                         assert.strictEqual(e.total, 0, 'if implementation can\'t compute event.total for gzipped content it is 0');</span></a>
<a name="20"><span class="lineNum">      20 </span>                :<span class="lineNoCov">          0 :                         assert(e.loaded &gt;= len, 'event.loaded should be set even if total is not computable');</span></a>
<a name="21"><span class="lineNum">      21 </span>                :<span class="lineNoCov">          0 :                 }</span></a>
<a name="22"><span class="lineNum">      22 </span>                :<span class="lineCov">          1 : </span></a>
<a name="23"><span class="lineNum">      23 </span>                :<span class="lineCov">          1 :                 xhr.onloadend = xhr.progress = null;</span></a>
<a name="24"><span class="lineNum">      24 </span>                :<span class="lineCov">          1 :         });</span></a>
<a name="25"><span class="lineNum">      25 </span>                :<span class="lineCov">          1 : </span></a>
<a name="26"><span class="lineNum">      26 </span>           [<span class="branchCov" title="Branch 0 was taken 4 times"> + </span>]:<span class="lineCov">          1 :         xhr.addEventListener('progress', (e) =&gt; {</span></a>
<a name="27"><span class="lineNum">      27 </span>           [<span class="branchCov" title="Branch 0 was taken 1 time"> + </span>]:<span class="lineCov">          4 :                 if(lastTotal === undefined){</span></a>
<a name="28"><span class="lineNum">      28 </span>                :<span class="lineCov">          1 :                         lastTotal = e.total;</span></a>
<a name="29"><span class="lineNum">      29 </span>                :<span class="lineCov">          1 :                 }</span></a>
<a name="30"><span class="lineNum">      30 </span>                :<span class="lineCov">          4 : </span></a>
<a name="31"><span class="lineNum">      31 </span>                :<span class="lineCov">          4 :                 if(e.lengthComputable &amp;&amp; e.total &amp;&amp; e.loaded){</span></a>
<a name="32"><span class="lineNum">      32 </span>                :<span class="lineCov">          4 :                         assert.strictEqual(e.total, lastTotal, 'event.total should remain invariant');</span></a>
<a name="33"><span class="lineNum">      33 </span>                :<span class="lineCov">          4 :                         assert(e.loaded &lt;= lastTotal, 'event.loaded should not exceed content-length');</span></a>
<a name="34"><span class="lineNum">      34 </span>           [<span class="branchNoCov" title="Branch 0 was not taken"> - </span>]:<span class="lineCov">          4 :                 }</span></a>
<a name="35"><span class="lineNum">      35 </span>                :<span class="lineNoCov">          0 :                 else{</span></a>
<a name="36"><span class="lineNum">      36 </span>                :<span class="lineNoCov">          0 :                         assert.strictEqual(e.total, 0, 'event.total should be 0');</span></a>
<a name="37"><span class="lineNum">      37 </span>                :<span class="lineNoCov">          0 :                 }</span></a>
<a name="38"><span class="lineNum">      38 </span>                :<span class="lineCov">          4 : </span></a>
<a name="39"><span class="lineNum">      39 </span>                :<span class="lineCov">          4 :                 assert(e.loaded &gt; lastLoaded, 'event.loaded should only ever increase');</span></a>
<a name="40"><span class="lineNum">      40 </span>                :<span class="lineCov">          4 :                 lastLoaded = e.loaded;</span></a>
<a name="41"><span class="lineNum">      41 </span>                :<span class="lineCov">          1 :         });</span></a>
<a name="42"><span class="lineNum">      42 </span>                :<span class="lineCov">          1 : </span></a>
<a name="43"><span class="lineNum">      43 </span>                :<span class="lineCov">          1 :         xhr.open('GET', `${activeURL}/image.gif?pipe=gzip|trickle(45000:d1:r2)`);</span></a>
<a name="44"><span class="lineNum">      44 </span>                :<span class="lineCov">          1 :         xhr.send();</span></a>
<a name="45"><span class="lineNum">      45 </span>                :<span class="lineCov">          1 : }</span></a>
<a name="46"><span class="lineNum">      46 </span>                :<span class="lineCov">          1 : </span></a>
<a name="47"><span class="lineNum">      47 </span>                :<span class="lineCov">          1 : /*</span></a>
<a name="48"><span class="lineNum">      48 </span>                :<span class="lineCov">          1 :  * progress-events-response-data-gzip.htm</span></a>
<a name="49"><span class="lineNum">      49 </span>                :<span class="lineCov">          1 :  *</span></a>
<a name="50"><span class="lineNum">      50 </span>                :<span class="lineCov">          1 : </span></a>
<a name="51"><span class="lineNum">      51 </span>                :<span class="lineCov">          1 : &lt;!doctype html&gt;</span></a>
<a name="52"><span class="lineNum">      52 </span>                :<span class="lineCov">          1 : &lt;html&gt;</span></a>
<a name="53"><span class="lineNum">      53 </span>                :<span class="lineCov">          1 :   &lt;head&gt;</span></a>
<a name="54"><span class="lineNum">      54 </span>                :<span class="lineCov">          1 :     &lt;title&gt;XMLHttpRequest: progress events and GZIP encoding&lt;/title&gt;</span></a>
<a name="55"><span class="lineNum">      55 </span>                :<span class="lineCov">          1 :     &lt;meta name=&quot;timeout&quot; content=&quot;long&quot;&gt;</span></a>
<a name="56"><span class="lineNum">      56 </span>                :<span class="lineCov">          1 :     &lt;script src=&quot;/resources/testharness.js&quot;&gt;&lt;/script&gt;</span></a>
<a name="57"><span class="lineNum">      57 </span>                :<span class="lineCov">          1 :     &lt;script src=&quot;/resources/testharnessreport.js&quot;&gt;&lt;/script&gt;</span></a>
<a name="58"><span class="lineNum">      58 </span>                :<span class="lineCov">          1 :     &lt;link rel=&quot;help&quot; href=&quot;https://xhr.spec.whatwg.org/#firing-events-using-the-progressevent-interface-for-http&quot; data-tested-assertations=&quot;following::p[contains(text(),'content-encodings')]&quot; /&gt;</span></a>
<a name="59"><span class="lineNum">      59 </span>                :<span class="lineCov">          1 :     &lt;!-- TODO: find better spec reference when https://www.w3.org/Bugs/Public/show_bug.cgi?id=25587 is fixed --&gt;</span></a>
<a name="60"><span class="lineNum">      60 </span>                :<span class="lineCov">          1 :   &lt;/head&gt;</span></a>
<a name="61"><span class="lineNum">      61 </span>                :<span class="lineCov">          1 :   &lt;body&gt;</span></a>
<a name="62"><span class="lineNum">      62 </span>                :<span class="lineCov">          1 :     &lt;div id=&quot;log&quot;&gt;&lt;/div&gt;</span></a>
<a name="63"><span class="lineNum">      63 </span>                :<span class="lineCov">          1 :     &lt;script&gt;</span></a>
<a name="64"><span class="lineNum">      64 </span>                :<span class="lineCov">          1 :       var test = async_test()</span></a>
<a name="65"><span class="lineNum">      65 </span>                :<span class="lineCov">          1 :       test.step(function() {</span></a>
<a name="66"><span class="lineNum">      66 </span>                :<span class="lineCov">          1 :         var client = new XMLHttpRequest()</span></a>
<a name="67"><span class="lineNum">      67 </span>                :<span class="lineCov">          1 :         *//*</span></a>
<a name="68"><span class="lineNum">      68 </span>                :<span class="lineCov">          1 : </span></a>
<a name="69"><span class="lineNum">      69 </span>                :<span class="lineCov">          1 :           Two behaviours are considered acceptable, so there are two ways to</span></a>
<a name="70"><span class="lineNum">      70 </span>                :<span class="lineCov">          1 :           pass this test</span></a>
<a name="71"><span class="lineNum">      71 </span>                :<span class="lineCov">          1 : </span></a>
<a name="72"><span class="lineNum">      72 </span>                :<span class="lineCov">          1 :           a) Set data for the compressed resource:</span></a>
<a name="73"><span class="lineNum">      73 </span>                :<span class="lineCov">          1 :             * event.total reflects the Content-length of the gzipp'ed resource</span></a>
<a name="74"><span class="lineNum">      74 </span>                :<span class="lineCov">          1 :             * event.loaded how many gzipped bytes have arrived over the wire so far</span></a>
<a name="75"><span class="lineNum">      75 </span>                :<span class="lineCov">          1 :             * lengthComputable is true</span></a>
<a name="76"><span class="lineNum">      76 </span>                :<span class="lineCov">          1 : </span></a>
<a name="77"><span class="lineNum">      77 </span>                :<span class="lineCov">          1 :           or</span></a>
<a name="78"><span class="lineNum">      78 </span>                :<span class="lineCov">          1 : </span></a>
<a name="79"><span class="lineNum">      79 </span>                :<span class="lineCov">          1 :           b) If the implementation does not provide progress details for the compressed</span></a>
<a name="80"><span class="lineNum">      80 </span>                :<span class="lineCov">          1 :           resource, set</span></a>
<a name="81"><span class="lineNum">      81 </span>                :<span class="lineCov">          1 :             * lengthComputable to false</span></a>
<a name="82"><span class="lineNum">      82 </span>                :<span class="lineCov">          1 :             * event.total to 0</span></a>
<a name="83"><span class="lineNum">      83 </span>                :<span class="lineCov">          1 :             * event.loaded to the number of bytes available so far after gzip decoding</span></a>
<a name="84"><span class="lineNum">      84 </span>                :<span class="lineCov">          1 : </span></a>
<a name="85"><span class="lineNum">      85 </span>                :<span class="lineCov">          1 :           Implications of this are tested here as follows:</span></a>
<a name="86"><span class="lineNum">      86 </span>                :<span class="lineCov">          1 : </span></a>
<a name="87"><span class="lineNum">      87 </span>                :<span class="lineCov">          1 :           * If lengthComputable is true:</span></a>
<a name="88"><span class="lineNum">      88 </span>                :<span class="lineCov">          1 :               * Event.total must match Content-length header</span></a>
<a name="89"><span class="lineNum">      89 </span>                :<span class="lineCov">          1 :               * event.loaded must only ever increase in progress events</span></a>
<a name="90"><span class="lineNum">      90 </span>                :<span class="lineCov">          1 :                 (and may never repeat its value).</span></a>
<a name="91"><span class="lineNum">      91 </span>                :<span class="lineCov">          1 :               * event.loaded must never exceed the Content-length.</span></a>
<a name="92"><span class="lineNum">      92 </span>                :<span class="lineCov">          1 : </span></a>
<a name="93"><span class="lineNum">      93 </span>                :<span class="lineCov">          1 :           * If lengthComputable is false:</span></a>
<a name="94"><span class="lineNum">      94 </span>                :<span class="lineCov">          1 :               * event.total should be 0</span></a>
<a name="95"><span class="lineNum">      95 </span>                :<span class="lineCov">          1 :               * event.loaded must only ever increase in progress events</span></a>
<a name="96"><span class="lineNum">      96 </span>                :<span class="lineCov">          1 :                 (and may never repeat its value).</span></a>
<a name="97"><span class="lineNum">      97 </span>                :<span class="lineCov">          1 :               * event.loaded should be the length of the decompressed content, i.e.</span></a>
<a name="98"><span class="lineNum">      98 </span>                :<span class="lineCov">          1 :                 bigger than Content-length header value when finished loading</span></a>
<a name="99"><span class="lineNum">      99 </span>                :<span class="lineCov">          1 : </span></a>
<a name="100"><span class="lineNum">     100 </span>                :<span class="lineCov">          1 :         *//*</span></a>
<a name="101"><span class="lineNum">     101 </span>                :<span class="lineCov">          1 :         var lastTotal;</span></a>
<a name="102"><span class="lineNum">     102 </span>                :<span class="lineCov">          1 :         var lastLoaded = -1;</span></a>
<a name="103"><span class="lineNum">     103 </span>                :<span class="lineCov">          1 :         client.addEventListener('loadend', test.step_func(function(e){</span></a>
<a name="104"><span class="lineNum">     104 </span>                :<span class="lineCov">          1 :           var len = parseInt(client.getResponseHeader('content-length'), 10)</span></a>
<a name="105"><span class="lineNum">     105 </span>                :<span class="lineCov">          1 :           if(e.lengthComputable){</span></a>
<a name="106"><span class="lineNum">     106 </span>                :<span class="lineCov">          1 :             assert_equals(e.total, len, 'event.total is content-length')</span></a>
<a name="107"><span class="lineNum">     107 </span>                :<span class="lineCov">          1 :             assert_equals(e.loaded, len, 'event.loaded should be content-length at loadend')</span></a>
<a name="108"><span class="lineNum">     108 </span>                :<span class="lineCov">          1 :           }else{</span></a>
<a name="109"><span class="lineNum">     109 </span>                :<span class="lineCov">          1 :             assert_equals(e.total, 0, 'if implementation can\'t compute event.total for gzipped content it is 0')</span></a>
<a name="110"><span class="lineNum">     110 </span>                :<span class="lineCov">          1 :             assert_true(e.loaded &gt;= len, 'event.loaded should be set even if total is not computable')</span></a>
<a name="111"><span class="lineNum">     111 </span>                :<span class="lineCov">          1 :           }</span></a>
<a name="112"><span class="lineNum">     112 </span>                :<span class="lineCov">          1 :           test.done();</span></a>
<a name="113"><span class="lineNum">     113 </span>                :<span class="lineCov">          1 :         }), false)</span></a>
<a name="114"><span class="lineNum">     114 </span>                :<span class="lineCov">          1 :         client.addEventListener('progress', test.step_func(function(e){</span></a>
<a name="115"><span class="lineNum">     115 </span>                :<span class="lineCov">          1 :           if(lastTotal === undefined){</span></a>
<a name="116"><span class="lineNum">     116 </span>                :<span class="lineCov">          1 :             lastTotal = e.total;</span></a>
<a name="117"><span class="lineNum">     117 </span>                :<span class="lineCov">          1 :           }</span></a>
<a name="118"><span class="lineNum">     118 </span>                :<span class="lineCov">          1 :           if(e.lengthComputable &amp;&amp; e.total &amp;&amp; e.loaded){</span></a>
<a name="119"><span class="lineNum">     119 </span>                :<span class="lineCov">          1 :             assert_equals(e.total, lastTotal, 'event.total should remain invariant')</span></a>
<a name="120"><span class="lineNum">     120 </span>                :<span class="lineCov">          1 :             assert_less_than_equal(e.loaded, lastTotal, 'event.loaded should not exceed content-length')</span></a>
<a name="121"><span class="lineNum">     121 </span>                :<span class="lineCov">          1 :           }else{</span></a>
<a name="122"><span class="lineNum">     122 </span>                :<span class="lineCov">          1 :             assert_equals(e.total, 0, 'event.total should be 0')</span></a>
<a name="123"><span class="lineNum">     123 </span>                :<span class="lineCov">          1 :           }</span></a>
<a name="124"><span class="lineNum">     124 </span>                :<span class="lineCov">          1 :           assert_greater_than(e.loaded, lastLoaded, 'event.loaded should only ever increase')</span></a>
<a name="125"><span class="lineNum">     125 </span>                :<span class="lineCov">          1 :           lastLoaded = e.loaded;</span></a>
<a name="126"><span class="lineNum">     126 </span>                :<span class="lineCov">          1 :         }), false)</span></a>
<a name="127"><span class="lineNum">     127 </span>                :<span class="lineCov">          1 :         // image.gif is 165375 bytes compressed. Sending 45000 bytes at a time with 1 second delay will load it in 4 seconds</span></a>
<a name="128"><span class="lineNum">     128 </span>                :<span class="lineCov">          1 :         client.open(&quot;GET&quot;, &quot;resources/image.gif?pipe=gzip|trickle(45000:d1:r2)&quot;, true)</span></a>
<a name="129"><span class="lineNum">     129 </span>                :<span class="lineCov">          1 :         client.send()</span></a>
<a name="130"><span class="lineNum">     130 </span>                :<span class="lineCov">          1 :       })</span></a>
<a name="131"><span class="lineNum">     131 </span>                :<span class="lineCov">          1 :     &lt;/script&gt;</span></a>
<a name="132"><span class="lineNum">     132 </span>                :<span class="lineCov">          1 :   &lt;/body&gt;</span></a>
<a name="133"><span class="lineNum">     133 </span>                :<span class="lineCov">          1 : &lt;/html&gt;</span></a>
<a name="134"><span class="lineNum">     134 </span>                :<span class="lineCov">          1 : </span></a>
<a name="135"><span class="lineNum">     135 </span>                :<span class="lineCov">          1 :  *</span></a>
<a name="136"><span class="lineNum">     136 </span>                :<span class="lineCov">          1 :  * progress-events-response-data-gzip.htm</span></a>
<a name="137"><span class="lineNum">     137 </span>                :<span class="lineCov">          1 :  */</span></a>
</pre>
      </td>
    </tr>
  </table>
  <br>

  <table width="100%" border=0 cellspacing=0 cellpadding=0>
    <tr><td class="ruler"><img src="../../../glass.png" width=3 height=3 alt=""></td></tr>
    <tr><td class="versionInfo">Generated by: <a href="http://ltp.sourceforge.net/coverage/lcov.php" target="_parent">LCOV version 1.14</a></td></tr>
  </table>
  <br>

</body>
</html>
